+++ vscode-lox/CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-[Unreleased]
+## [Unreleased]
+
+### Fixed
+
+- Stop showing "this" from nested classes when finding references of "this".
+
+### Changed
+
+- Improve resolving of object properties to their definitions.
 
 ## [4.0.2] - 2025-12-26